refractor Commands Sample
