+++ src/Avalonia.Samples/MVVM/CommandSample/App.axaml.cs
@@ -19,7 +19,7 @@ namespace CommandSample
             {
                 desktop.MainWindow = new MainWindow
                 {
-                    DataContext = new MainWindowViewModel(),
+                    DataContext = new MainViewModel(),
                 };
             }
