+++ src/Avalonia.Samples/MVVM/CommandSample/ViewModels/MainViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace CommandSample.ViewModels;
+
+public class MainViewModel : ViewModelBase
+{
+    public ReactiveUiCommandsViewModel ReactiveUiCommandsViewModel { get; } = new ReactiveUiCommandsViewModel();
+    
+    public CommunityToolkitCommandsViewModel CommunityToolkitCommandsViewModel { get; } = new CommunityToolkitCommandsViewModel();
+}
