+++ src/Avalonia.Samples/MVVM/CommandSample/Views/CommunityToolkitCommandsSampleView.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace CommandSample.Views;
+
+public partial class CommunityToolkitCommandsSampleView : UserControl
+{
+    public CommunityToolkitCommandsSampleView()
+    {
+        InitializeComponent();
+    }
+}
