+++ src/Avalonia.Samples/MVVM/CommandSample/Views/ReactiveUiCommandsSampleView.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace CommandSample.Views;
+
+public partial class ReactiveUiCommandsSampleView : UserControl
+{
+    public ReactiveUiCommandsSampleView()
+    {
+        InitializeComponent();
+    }
+}
